[CPU] DetectionOutput dynamism enabling: test movement
