+++ inference-engine/tests/functional/plugin/cpu/shared_tests_instances/single_layer_tests/detection_output.cpp
@@ -3,11 +3,23 @@
 //
 
 #include "single_layer_tests/detection_output.hpp"
-#include "single_layer_tests/detection_output_attributes.hpp"
 
 using namespace LayerTestsDefinitions;
 
 namespace {
+
+const int numClasses = 11;
+const int backgroundLabelId = 0;
+const std::vector<int> topK = {75};
+const std::vector<std::vector<int>> keepTopK = { {50}, {100} };
+const std::vector<std::string> codeType = {"caffe.PriorBoxParameter.CORNER", "caffe.PriorBoxParameter.CENTER_SIZE"};
+const float nmsThreshold = 0.5f;
+const float confidenceThreshold = 0.3f;
+const std::vector<bool> clipAfterNms = {true, false};
+const std::vector<bool> clipBeforeNms = {true, false};
+const std::vector<bool> decreaseLabelId = {true, false};
+const float objectnessScore = 0.4f;
+const std::vector<size_t> numberBatch = {1, 2};
 
 const auto commonAttributes = ::testing::Combine(
         ::testing::Values(numClasses),
